+++ examples/simulate_fmi2_me.c
@@ -87,7 +87,7 @@ int main(int argc, char* argv[]) {
 
     CALL(recordVariables(S, outputFile));
 
-    size_t step = 0;
+    uint64_t step = 0;
 
     while (!eventInfo.terminateSimulation) {
